requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
disnake==2.4.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
multidict==6.0.2
mutagen==1.45.1
pycparser==2.21
pycryptodomex==3.14.1
PyNaCl==1.4.0
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.32
websockets==10.2
wincertstore==0.2
yarl==1.7.2
yt-dlp==2022.3.8.2
